cleanup

+++ tegbp.hpp
@@ -32,10 +32,10 @@ struct mem_pool{
     double * node;
     int * sae ;
     double * v_norms;
-    int16 *indices;
+    int32 *indices;
     int *timestamps;
-    int16 H;
-    int16 W;
+    int32 H;
+    int32 W;
     int B;
     std::string data_name;
     int WINSIZE;
@@ -46,8 +46,8 @@ void debug_output(mem_pool pool);
 void save_data(mem_pool pool, int seq_id, int index, int c_time);
 void process_batch(mem_pool pool, int b_ptr);
 mem_pool initialize(mem_pool pool);
-int sub2ind(int16 x, int16 y, int node_index, int dir, int16 H, int16 W);
-int sub2ind_obs(int16 x, int16 y, int16 H, int16 W);
-int sub2ind_slf(int16 x, int16 y, int16 H, int16 W);
+int sub2ind(int32 x, int32 y, int node_index, int dir, int32 H, int32 W);
+int sub2ind_obs(int32 x, int32 y, int32 H, int32 W);
+int sub2ind_slf(int32 x, int32 y, int32 H, int32 W);
 // void init_sae(mem_pool pool);
 #endif
